% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{backtrack_descent}
\alias{backtrack_descent}
\title{Backtracking for steepest descent}
\usage{
backtrack_descent(fx, x, t, df, d, alpha = 0.5, beta = 0.9)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{x}{current parameter estimate}

\item{t}{current step-size}

\item{df}{the value of the gradient of objective function evaluated at the current x}

\item{d}{descent direction vector}

\item{alpha}{the backtracking parameter}

\item{beta}{the decrementing multiplier}
}
\description{
Backtracking for steepest descent
}
