% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{fx_logistic}
\alias{fx_logistic}
\title{Objective Function for Logistic Regression}
\usage{
fx_logistic(y, X, beta, lambda = 0)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{lambda}{regularization parameter}
}
\description{
Objective Function for Logistic Regression
}
