% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{gradf_logistic}
\alias{gradf_logistic}
\title{Gradient for Logistic Regession}
\usage{
gradf_logistic(y, X, beta, lambda = 0)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{lambda}{regularization parameter}
}
\description{
Gradient for Logistic Regession
}
