% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{gradient_step}
\alias{gradient_step}
\title{Gradient Step}
\usage{
gradient_step(gradf, t, x)
}
\arguments{
\item{gradf}{handle to function that returns gradient of objective function}

\item{t}{step-size}

\item{x}{current parameter estimate}
}
\description{
Gradient Step
}
