% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave One Out}
\usage{
leave_one_out(y, X, lambda)
}
\arguments{
\item{y}{response variables}

\item{X}{design matrix}

\item{lambda}{vector of tuning parameters}
}
\description{
\code{leave_one_out} returns the leave-one-out
for a sequence of regularization parameter values.
}
