% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{logistic_ridge_newton}
\alias{logistic_ridge_newton}
\title{Damped Newton's Method for Fitting Ridge Logistic Regression}
\usage{
logistic_ridge_newton(X, y, beta, lambda = 0, naive = TRUE,
  max_iter = 100, tol = 0.001)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Binary response}

\item{beta}{Initial regression coefficient vector}

\item{lambda}{regularization parameter}

\item{naive}{Boolean variable; TRUE if using Cholesky on the Hessian}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Damped Newton's Method for Fitting Ridge Logistic Regression
}
