% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{newton_step_naive}
\alias{newton_step_naive}
\title{Compute Newton Step (Naive) for logistic ridge regression}
\usage{
newton_step_naive(X, y, beta, g, lambda)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Binary response vector}

\item{beta}{Current regression vector estimate}

\item{g}{Gradient vector}

\item{lambda}{Regularization parameter}
}
\description{
Compute Newton Step (Naive) for logistic ridge regression
}
