% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{power_method_dense}
\alias{power_method_dense}
\title{Power Method for dense matrices}
\usage{
power_method_dense(A, x, max_iter, tol)
}
\arguments{
\item{A}{The input matrix}

\item{x}{initial guess}

\item{max_iter}{maximum number of iterations}

\item{tol}{relative error stopping criterion}
}
\description{
\code{power_method_dense} applies the power method to estimate
the eigenvector of a dense matrix associated with its largest eigenvector.
}
