% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{power_method_low_rank}
\alias{power_method_low_rank}
\title{Power Method for low rank matrices}
\usage{
power_method_low_rank(U, V, x, max_iter, tol)
}
\arguments{
\item{U}{The left input factor matrix}

\item{V}{The right input factor matrix}

\item{x}{initial guess}

\item{max_iter}{maximum number of iterations}

\item{tol}{relative error stopping criterion}
}
\description{
\code{power_method_low_rank} applies the power method to estimate
the eigenvector of a low rank matrix associated with its largest eigenvector.
}
