% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{power_method_sparse}
\alias{power_method_sparse}
\title{Power Method for sparse matrices}
\usage{
power_method_sparse(A, x, max_iter, tol)
}
\arguments{
\item{A}{The input matrix}

\item{x}{initial guess}

\item{max_iter}{maximum number of iterations}

\item{tol}{relative error stopping criterion}
}
\description{
\code{power_method_sparse} applies the power method to estimate
the eigenvector of a sparse matrix associated with its largest eigenvector.
}
